+++ packages/common/src/groups/_internal/GroupUiSchemaCreateFollowers.ts
@@ -108,8 +108,8 @@ export const buildUiSchema = async (
                 options: {
                   control: "hub-field-input-radio",
                   labels: [
-                    `{{${i18nScope}.fields.membershipAccess.org:translate}}`,
-                    `{{${i18nScope}.fields.membershipAccess.collab:translate}}`,
+                    `{{${i18nScope}.fields.membershipAccess.org.description:translate}}`,
+                    `{{${i18nScope}.fields.membershipAccess.collab.description:translate}}`,
                     `{{${i18nScope}.fields.membershipAccess.createFollowers.any:translate}}`,
                   ],
                   disabled: [false, false, entity.isSharedUpdate],
@@ -122,7 +122,7 @@ export const buildUiSchema = async (
                 options: {
                   control: "hub-field-input-radio",
                   labels: [
-                    `{{${i18nScope}.fields.contributeContent.all:translate}}`,
+                    `{{${i18nScope}.fields.contributeContent.members.description:translate}}`,
                     `{{${i18nScope}.fields.contributeContent.createFollowers.admins:translate}}`,
                   ],
                 },